chore: cleaned up client files
